% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbResume.R
\name{playoff_picture}
\alias{playoff_picture}
\title{Select College Football Playoff Teams}
\usage{
playoff_picture(
  df,
  n = 12,
  offenseWeight = 0.1,
  defenseWeight = 0.1,
  controlWeight = 0.05,
  penaltyWeight = -0.05,
  rankedWeight = 0.5,
  regWeight = 0.15,
  confWeight = 0.05
)
}
\arguments{
\item{df}{A df with features: pass_yards_diff_pg, rush_yards_diff_pg, td_diff_pg,
interception_diff_pg, fumble_diff_pg, sack_diff_pg, tfl_diff_pg,
turnover_diff_pg, possession_diff, penalties_pg, penalty_yards_pg,
Ranked_WinPct, Conf_WinPct, RegSeason_WinPct, along with Team, Conference,
and Record}

\item{n}{Number of teams to select (default 12)}

\item{offenseWeight}{Weighting of offensive performance for resume score}

\item{defenseWeight}{Weighting of defensive performance for resume score}

\item{controlWeight}{Weighting of control performance for resume score}

\item{penaltyWeight}{Weighting of penalties for resume score}

\item{rankedWeight}{Weighting of WinPct against ranked opponents for resume score}

\item{regWeight}{Weighting of WinPct in the regular season for resume score}

\item{confWeight}{Weighting of WinPct against conference opponents for resume score}
}
\value{
Returns the top n teams based on resume scores
}
\description{
Evaluate each CFB team and return the top n teams for the CFB playoff
}
